requirements.txt
requests==2.32.5
beautifulsoup4==4.12.3
lxml>=6.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1
